Flatten RaiseAccumulate in arrow.core.raise
