+++ arrow-libs/core/arrow-core/src/commonMain/kotlin/arrow/core/Iterable.kt
@@ -4,7 +4,7 @@ package arrow.core
 
 import arrow.core.Either.Left
 import arrow.core.Either.Right
-import arrow.core.raise.Raise
+import arrow.core.raise.RaiseAccumulate
 import arrow.core.raise.fold
 import arrow.typeclasses.Monoid
 import arrow.typeclasses.Semigroup
@@ -491,13 +491,13 @@ public fun <A> Iterable<A?>.sequenceNullable(): List<A>? =
 @OptIn(ExperimentalTypeInference::class)
 public inline fun <Error, A, B> Iterable<A>.mapOrAccumulate(
   combine: (Error, Error) -> Error,
-  @BuilderInference transform: AccumulatingRaise<Error>.(A) -> B,
+  @BuilderInference transform: RaiseAccumulate<Error>.(A) -> B,
 ): Either<Error, List<B>> {
   var left: Any? = EmptyValue
   val right = ArrayList<B>(collectionSizeOrDefault(10))
   for (item in this)
     fold(
-      { transform(AccumulatingRaise(this), item) },
+      { transform(RaiseAccumulate(this), item) },
       { errors -> left = EmptyValue.combine(left, errors.reduce(combine), combine) },
       { b -> right.add(b) }
     )
@@ -511,12 +511,12 @@ public inline fun <Error, A, B> Iterable<A>.mapOrAccumulate(
  */
 @OptIn(ExperimentalTypeInference::class)
 public inline fun <Error, A, B> Iterable<A>.mapOrAccumulate(
-  @BuilderInference transform: AccumulatingRaise<Error>.(A) -> B,
+  @BuilderInference transform: RaiseAccumulate<Error>.(A) -> B,
 ): Either<NonEmptyList<Error>, List<B>> {
   val left = mutableListOf<Error>()
   val right = ArrayList<B>(collectionSizeOrDefault(10))
   for (item in this)
-    fold({ transform(AccumulatingRaise(this), item) }, { error -> left.addAll(error) }, { b -> right.add(b) })
+    fold({ transform(RaiseAccumulate(this), item) }, { error -> left.addAll(error) }, { b -> right.add(b) })
   return left.toNonEmptyListOrNull()?.left() ?: right.right()
 }
