+++ arrow-libs/core/arrow-core/src/commonMain/kotlin/arrow/core/Sequence.kt
@@ -4,13 +4,11 @@ package arrow.core
 
 import arrow.core.Either.Left
 import arrow.core.Either.Right
-import arrow.core.raise.Raise
-import arrow.core.raise.either
+import arrow.core.raise.RaiseAccumulate
 import arrow.core.raise.fold
 import arrow.typeclasses.Monoid
 import arrow.typeclasses.Semigroup
 import kotlin.experimental.ExperimentalTypeInference
-import kotlin.jvm.JvmName
 
 public fun <B, C, D, E> Sequence<B>.zip(
   c: Sequence<C>,
@@ -749,22 +747,22 @@ public fun <E, A, B> Sequence<A>.traverse(
 
 public fun <Error, A, B> Sequence<A>.mapOrAccumulate(
   combine: (Error, Error) -> Error,
-  @BuilderInference transform: AccumulatingRaise<Error>.(A) -> B
+  @BuilderInference transform: RaiseAccumulate<Error>.(A) -> B
 ): Either<Error, List<B>> {
   var left: Any? = EmptyValue
   val right = mutableListOf<B>()
   for (item in this)
-    fold({ transform(AccumulatingRaise(this), item) }, { errors -> left = EmptyValue.combine(left, errors.reduce(combine), combine) }, { b -> right.add(b) })
+    fold({ transform(RaiseAccumulate(this), item) }, { errors -> left = EmptyValue.combine(left, errors.reduce(combine), combine) }, { b -> right.add(b) })
   return if (left !== EmptyValue) EmptyValue.unbox<Error>(left).left() else right.right()
 }
 
 public fun <Error, A, B> Sequence<A>.mapOrAccumulate(
-  @BuilderInference transform: AccumulatingRaise<Error>.(A) -> B
+  @BuilderInference transform: RaiseAccumulate<Error>.(A) -> B
 ): Either<NonEmptyList<Error>, List<B>> {
   val left = mutableListOf<Error>()
   val right = mutableListOf<B>()
   for (item in this)
-    fold({ transform(AccumulatingRaise(this), item) }, { errors -> left.addAll(errors) }, { b -> right.add(b) })
+    fold({ transform(RaiseAccumulate(this), item) }, { errors -> left.addAll(errors) }, { b -> right.add(b) })
   return left.toNonEmptyListOrNull()?.left() ?: right.right()
 }
