+++ arrow-libs/fx/arrow-fx-coroutines/src/commonMain/kotlin/arrow/fx/coroutines/ParMap.kt
@@ -1,6 +1,6 @@
 package arrow.fx.coroutines
 
-import arrow.core.AccumulatingRaise
+import arrow.core.raise.RaiseAccumulate
 import arrow.core.Either
 import arrow.core.NonEmptyList
 import arrow.core.flattenOrAccumulate
@@ -34,16 +34,16 @@ public suspend fun <A, B> Iterable<A>.parMap(
 }
 
 /** Temporary intersection type, until we have context receivers */
-public class ScopedAccumulatingRaise<Error>(
+public class ScopedRaiseAccumulate<Error>(
   raise: Raise<NonEmptyList<Error>>,
   scope: CoroutineScope
-) : CoroutineScope by scope, AccumulatingRaise<Error>(raise)
+) : CoroutineScope by scope, RaiseAccumulate<Error>(raise)
 
 public suspend fun <Error, A, B> Iterable<A>.parMapOrAccumulate(
   context: CoroutineContext = EmptyCoroutineContext,
   concurrency: Int,
   combine: (Error, Error) -> Error,
-  transform: suspend ScopedAccumulatingRaise<Error>.(A) -> B
+  transform: suspend ScopedRaiseAccumulate<Error>.(A) -> B
 ): Either<Error, List<B>> =
   coroutineScope {
     val semaphore = Semaphore(concurrency)
@@ -51,7 +51,7 @@ public suspend fun <Error, A, B> Iterable<A>.parMapOrAccumulate(
       async(context) {
         either {
           semaphore.withPermit {
-            transform(ScopedAccumulatingRaise(this, this@coroutineScope), it)
+            transform(ScopedRaiseAccumulate(this, this@coroutineScope), it)
           }
         }
       }
@@ -61,13 +61,13 @@ public suspend fun <Error, A, B> Iterable<A>.parMapOrAccumulate(
 public suspend fun <Error, A, B> Iterable<A>.parMapOrAccumulate(
   context: CoroutineContext = EmptyCoroutineContext,
   combine: (Error, Error) -> Error,
-  transform: suspend ScopedAccumulatingRaise<Error>.(A) -> B
+  transform: suspend ScopedRaiseAccumulate<Error>.(A) -> B
 ): Either<Error, List<B>> =
   coroutineScope {
     map {
       async(context) {
         either {
-          transform(ScopedAccumulatingRaise(this, this@coroutineScope), it)
+          transform(ScopedRaiseAccumulate(this, this@coroutineScope), it)
         }
       }
     }.awaitAll().flattenOrAccumulate(combine)
@@ -76,7 +76,7 @@ public suspend fun <Error, A, B> Iterable<A>.parMapOrAccumulate(
 public suspend fun <Error, A, B> Iterable<A>.parMapOrAccumulate(
   context: CoroutineContext = EmptyCoroutineContext,
   concurrency: Int,
-  transform: suspend ScopedAccumulatingRaise<Error>.(A) -> B
+  transform: suspend ScopedRaiseAccumulate<Error>.(A) -> B
 ): Either<NonEmptyList<Error>, List<B>> =
   coroutineScope {
     val semaphore = Semaphore(concurrency)
@@ -84,7 +84,7 @@ public suspend fun <Error, A, B> Iterable<A>.parMapOrAccumulate(
       async(context) {
         either {
           semaphore.withPermit {
-            transform(ScopedAccumulatingRaise(this, this@coroutineScope), it)
+            transform(ScopedRaiseAccumulate(this, this@coroutineScope), it)
           }
         }
       }
@@ -93,13 +93,13 @@ public suspend fun <Error, A, B> Iterable<A>.parMapOrAccumulate(
 
 public suspend fun <Error, A, B> Iterable<A>.parMapOrAccumulate(
   context: CoroutineContext = EmptyCoroutineContext,
-  transform: suspend ScopedAccumulatingRaise<Error>.(A) -> B
+  transform: suspend ScopedRaiseAccumulate<Error>.(A) -> B
 ): Either<NonEmptyList<Error>, List<B>> =
   coroutineScope {
     map {
       async(context) {
         either {
-          transform(ScopedAccumulatingRaise(this, this@coroutineScope), it)
+          transform(ScopedRaiseAccumulate(this, this@coroutineScope), it)
         }
       }
     }.awaitAll().flattenOrAccumulate()
